+++ src/app/notification/notification.demo.ts
@@ -12,6 +12,7 @@ import { SohoNotificationService } from 'ids-enterprise-ng';
 })
 export class NotificationDemoComponent implements OnInit {
   private counter = 0;
+  private current = 0;
   // Inject Notification Servivce as a dependency and get an instance variable
   constructor(private notificationService: SohoNotificationService) { }
   ngOnInit() { }
@@ -20,9 +21,33 @@ export class NotificationDemoComponent implements OnInit {
     this.notificationService.show({ id: `notif-${this.counter++}`, parent: '.page-container.scrollable', message: `${this.counter}) This is a Toast message`, type });
   }
 
-  hideNotification() {
+  hideFirstNotification() {
     if (this.counter > 0) {
-      this.notificationService.hide(`notif-${--this.counter}`);
+      this.notificationService.hide(`notif-${this.current++}`);
+      this.counter--;
+
+      if (this.counter === 0) {
+        this.current = 0;
+      }
+    }
+  }
+
+  hideLatestNotification() {
+    if (this.counter > 0) {
+      this.notificationService.hideLatest();
+      this.counter--;
+
+      if (this.counter === 0) {
+        this.current = 0;
+      }
+    }
+  }
+
+  hideAllNotification() {
+    if (this.counter > 0) {
+      this.notificationService.hideAll();
+      this.counter = 0;
+      this.current = 0;
     }
   }
 }